+++ CHANGELOG.md
@@ -27,10 +27,12 @@ Version 1.2.2 (not released yet)
 - fix decal fade out effect (RF bug)
 - fix invalid player entity model when undercover (RF bug)
 - speed up server list refreshing
-- add linear_pitch command for making pitch control linear
+- add linear_pitch command for making pitch control linear (configuration is persisted)
 - add 'findlevel' command as alias for 'findmap'
 - add 'map' command as alias for 'level'
 - allow 'level' command outside of multiplayer game and remove now redundant 'levelsp' command
+- add 'show_enemy_bullets' command for toggling enemy bullets visibility (configuration is persisted) - it was forced on before
+- add persisting of volumetric lights (glares) configuration changed by 'vli' command
 - fix multiple security vulnerabilities
 - stability improvements